add - register route and tests

+++ Back-Far-Away-Users/src/main/java/be/d2l/service/UserService.java
@@ -1,6 +1,8 @@
 package be.d2l.service;
 
 import be.d2l.Exceptions.UnauthorizedException;
+import be.d2l.Exceptions.UserAlreadyExistsException;
+import be.d2l.Exceptions.UserNotFoundException;
 import be.d2l.config.CustomProperties;
 import be.d2l.model.User;
 import be.d2l.repo.UserRepository;
@@ -29,17 +31,19 @@ public class UserService {
         return repo.findAll();
     }
 
-    public void deleteUser(int id){
+    public void deleteUser(int id) throws UserNotFoundException {
+        if (!repo.existsById(id)) throw new UserNotFoundException("No user found with id " + id);
         repo.deleteById(id);
     }
 
-    public User createUser(User receivedUser){
+    public User createUser(User receivedUser) throws UserAlreadyExistsException {
+        if (repo.existsByMail(receivedUser.getMail())) throw new UserAlreadyExistsException();
         User u = new User();
         return repo.save(setUser(u,receivedUser));
     }
 
-    public User updateUser(User receivedUser, int id){
-        User u = repo.findById(id).orElseThrow(InternalError::new);
+    public User updateUser(User receivedUser, int id) throws UserNotFoundException {
+        User u = repo.findById(id).orElseThrow(() -> new UserNotFoundException("User with id " + id + " not found"));
         return repo.save(setUser(u,receivedUser));
     }
 
@@ -52,19 +56,26 @@ public class UserService {
         if (userFound == null)
              throw new UnauthorizedException("Email incorrect");
         BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
-        if(encoder.matches(userFound.getPassword(), encoder.encode(password)))
+        if(!encoder.matches(password, userFound.getPassword()))
             throw new UnauthorizedException("Password incorrect");
         return userFound;
     }
 
     private User setUser(User u,User receivedUser){
-        u.setSurname(receivedUser.getSurname());
-        u.setFirstName(receivedUser.getFirstName());
-        u.setNickname(receivedUser.getFirstName());
-        u.setBirthDate(receivedUser.getBirthDate());
-        u.setAddress(receivedUser.getAddress());
-        u.setMail(receivedUser.getMail());
-        u.setPassword(new BCryptPasswordEncoder().encode(receivedUser.getPassword()));
+        if (receivedUser.getSurname() != null)
+            u.setSurname(receivedUser.getSurname());
+        if (receivedUser.getFirstName() != null)
+            u.setFirstName(receivedUser.getFirstName());
+        if (receivedUser.getNickname() != null)
+            u.setNickname(receivedUser.getNickname());
+        if (receivedUser.getBirthDate() != null)
+            u.setBirthDate(receivedUser.getBirthDate());
+        if (receivedUser.getAddress() != null)
+            u.setAddress(receivedUser.getAddress());
+        if (receivedUser.getMail() != null)
+            u.setMail(receivedUser.getMail());
+        if (receivedUser.getPassword() != null)
+            u.setPassword(new BCryptPasswordEncoder().encode(receivedUser.getPassword()));
         return u;
     }
